app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
Pillow>=9.0.0
numpy>=1.26.0
tensorflow>=2.16.0
python-dotenv>=1.0.0
pydantic>=2.7.0
